+++ src/index.js
@@ -1,5 +1,5 @@
-export {getAltInstance} from './altInstance';
-export {setAltInstance} from './altInstance';
+export { getAltInstance } from './altInstance';
+export { setAltInstance } from './altInstance';
 
 export connect from './decorators/connect';
 export bindCalls from './decorators/bindCalls';
@@ -7,13 +7,17 @@ export bindHandlers from './decorators/bindHandlers';
 export bindActions from './decorators/bindActions';
 
 export ImmutableStore from './ImmutableStore';
-export {resetStores} from './ImmutableStore';
+export { resetStores } from './ImmutableStore';
 
-export {connectAlternative} from './decorators/connect';
+export { connectAlternative } from './decorators/connect';
 export callSeries from './callSeries';
 export createStore from './createStore';
 export callFactory from './callFactory';
+export createAction from './createAction';
 export createActions from './createActions';
 export handlerFactory from './handlerFactory';
 export getSources from './getSources';
 
+export ActionListeners from 'alt-utils/lib/ActionListeners';
+export { decorate, datasource, bind } from 'alt-utils/lib/decorators';
+export immutable from 'alt-utils/lib/ImmutableUtil';